Initial Implementation of Auto Moderation
This version comes with 4 base actions: 'reject', 'edit', 'ban', and
'delete'. Reject will reject the post outright. Edit allows the auto
moderation to edit the post using either a template or a regex. Ban will
allow the post but ban the user. Delete will also allow the post but set
it to 'deleted' by default so that it only shows up to mods and admin.
Any of these actions can be combined in one rule. Although reject only
makes sense with ban.

Each options, except delete, has some options that may be set that allow
either modifying the post in some way, returning data to the users, or
setting the limit on the ban time period.

Each rule can have any number of conditions but they all must be met for
the rule to enable the action.

Rules must have a name and an optional description. There is no sorting
order.